- Removed basefile.txt writes - CollisionHeader can now be defined in the XML - Fixed nonmatching issue with ZCutscene - LOD Limbs should work, but still needs testing - Fixed bug where default limb DL names would not be generated - CFG Files no longer generated, since it is no longer necessary.
