+++ ZAPD/ZAPD.vcxproj.filters
@@ -189,6 +189,9 @@
     <ClCompile Include="ZRoom\Commands\ZRoomCommandUnk.cpp">
       <Filter>Source Files\Z64\ZRoom\Commands</Filter>
     </ClCompile>
+    <ClCompile Include="ZCollision.cpp">
+      <Filter>Source Files\Z64</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Path.h">
@@ -403,6 +406,9 @@
     </ClInclude>
     <ClInclude Include="sqlite_orm.h">
       <Filter>Header Files\Libraries</Filter>
+    </ClInclude>
+    <ClInclude Include="ZCollision.h">
+      <Filter>Header Files\Z64</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
